Add single bucket connection type

+++ Apps.AmazonS3/Actions/BucketActions.cs
@@ -1,8 +1,10 @@
 ﻿using Amazon.S3.Model;
+using Apps.AmazonS3.Constants;
 using Apps.AmazonS3.Models.Request;
 using Apps.AmazonS3.Models.Response;
 using Blackbird.Applications.Sdk.Common;
 using Blackbird.Applications.Sdk.Common.Actions;
+using Blackbird.Applications.Sdk.Common.Exceptions;
 using Blackbird.Applications.Sdk.Common.Invocation;
 
 namespace Apps.AmazonS3.Actions;
@@ -13,6 +15,9 @@ public class BucketActions(InvocationContext invocationContext) : AmazonInvocabl
     [Action("Create bucket", Description = "Create an S3 bucket.")]
     public async Task<BucketResponse> CreateBucket([ActionParameter] [Display("Bucket name")] string bucketName)
     {
+        if (CurrentConnectionType != ConnectionTypes.AllBuckets)
+            throw new PluginMisconfigurationException($"Currently selected connection supports only '{ConnectedBucket}' bucket. Please, switch to 'All buckets' conection for working with buckets themselves.");
+
         var request = new PutBucketRequest
         {
             BucketName = bucketName,
@@ -27,12 +32,15 @@ public class BucketActions(InvocationContext invocationContext) : AmazonInvocabl
     [Action("Delete bucket", Description = "Delete an S3 bucket.")]
     public async Task DeleteBucket([ActionParameter] BucketRequest bucket)
     {
+        if (CurrentConnectionType != ConnectionTypes.AllBuckets)
+            throw new PluginMisconfigurationException($"Currently selected connection supports only '{ConnectedBucket}' bucket. Please, switch to 'All buckets' conection for working with buckets themselves.");
+
         var request = new DeleteBucketRequest
         {
-            BucketName = bucket.BucketName,
+            BucketName = bucket.BucketName!,
         };
 
-        var client = await CreateBucketClient(bucket.BucketName);
+        var client = await CreateBucketClient(bucket.BucketName!);
         await ExecuteAction(() => client.DeleteBucketAsync(request));
     }
 }
